+++ test/test_helper.exs
@@ -9,4 +9,5 @@ ExUnit.start(exclude: postgresql_exclusions)
 
 Code.require_file("support/postgresql_case.ex", __DIR__)
 Code.require_file("support/telemetry_handler.ex", __DIR__)
+Code.require_file("support/postgresql_response_proxy.ex", __DIR__)
 Code.require_file("support/worker_tasks.ex", __DIR__)